all ints?

+++ scripts/campus.py
@@ -59,7 +59,7 @@ class campus:
                 for i, sensor in enumerate(room_object.sensors):
                     sensor_object = room_object.sensors[sensor]
                     deviation = sensor_object.getOlder(index)
-                    entry = [int(epoch_time), int(building_object.id), int(room_object.id), int(sensor_object.id), np.int64(deviation)]
+                    entry = [np.int64(epoch_time), np.int64(building_object.id), np.int64(room_object.id), int(sensor_object.id), np.int64(deviation)]
                     self.metrics['temperature'].append(entry)
 
             for corridor in building_object.corridors:
@@ -68,7 +68,7 @@ class campus:
                     sensor_object = corridor_object.sensors[sensor]
 
                     deviation = sensor_object.getOlder(index)
-                    entry = [int(epoch_time), int(building_object.id), int(corridor_object.id), int(sensor_object.id), np.int64(deviation)]
+                    entry = [np.int64(epoch_time), np.int64(building_object.id), np.int64(corridor_object.id), np.int64(sensor_object.id), np.int64(deviation)]
                     self.metrics['temperature'].append(entry)
 
     def makePipe(self):
