+++ scripts/corridor.py
@@ -23,7 +23,7 @@ class corridor:
                 self.sensors[i] = sensor('thermostat', numRows, sensor_attributes)
                 self.id = attrs['id']
         else:
-            self.id = str(uuid.uuid4().int)[:7]
+            self.id = np.int64(str(uuid.uuid4().int)[:7])
 
             num_thermostat = rd.randint(1, 2)
             num_co2 = rd.randint(1, 3)
